Fixed Android navigation and CSS, fully implemented changes from last commit on master
The last commit on master completely broke the Android version of the app and changes were made to html without concern for Android compatibility. This branch forked off before the "Adding Navigation for iOS + Simplification of code" commit on master. Now, in this commit, we have all the positive changes made to the index.js and simplification of css between platforms while maintaining the full intended functionality of the iOS and Android platforms so far.